# services/api-ts/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Usuários",
    "version": "1.0.0",
    "description": "Documentação da API com Swagger (via JSON separado)"
  },
  "servers": [
    {
      "url": "http://localhost:4500",
      "description": "Servidor local"
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "summary": "Listar todos os usuários",
        "tags": ["Usuários"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de usuários retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/User" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar um novo usuário",
        "tags": ["Usuários"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "400": {
            "description": "Erro de validação ou dados inválidos"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Buscar usuário por ID",
        "tags": ["Usuários"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      },
      "put": {
        "summary": "Atualizar usuário por ID",
        "tags": ["Usuários"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "400": {
            "description": "Erro de validação"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      },
      "delete": {
        "summary": "Remover usuário por ID",
        "tags": ["Usuários"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário removido com sucesso"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login do usuário e geração de token",
        "tags": ["Autenticação"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token gerado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOi..."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": { "type": "string", "example": "60f8b4f7c2a5e1234567890a" },
          "nome": { "type": "string", "example": "João" },
          "email": { "type": "string", "example": "joao@email.com" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "UserInput": {
        "type": "object",
        "properties": {
          "nome": { "type": "string", "example": "João" },
          "email": { "type": "string", "example": "joao@email.com" },
          "senha": { "type": "string", "example": "123456" }
        },
        "required": ["nome", "email", "senha"]
      },
      "LoginInput": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "example": "joao@email.com" },
          "senha": { "type": "string", "example": "123456" }
        },
        "required": ["email", "senha"]
      }
    }
  }
}